+++ coloredlogs/syslog.py
@@ -1,7 +1,7 @@
 # Easy to use system logging for Python's logging module.
 #
 # Author: Peter Odding <peter@peterodding.com>
-# Last Change: April 17, 2017
+# Last Change: December 10, 2020
 # URL: https://coloredlogs.readthedocs.io
 
 """
@@ -31,6 +31,10 @@ import logging.handlers
 import os
 import socket
 import sys
+
+# External dependencies.
+from humanfriendly import coerce_boolean
+from humanfriendly.compat import on_macos, on_windows
 
 # Modules included in our package.
 from coloredlogs import (
@@ -126,9 +130,16 @@ def enable_system_logging(programname=None, fmt=None, logger=None, reconfigure=T
               connection to the system logging daemon fails :data:`None` is
               returned.
 
+    As of release 15.0 this function uses :func:`is_syslog_supported()` to
+    check whether system logging is supported and appropriate before it's
+    enabled.
+
     .. note:: When the logger's effective level is too restrictive it is
               relaxed (refer to `notes about log levels`_ for details).
     """
+    # Check whether system logging is supported / appropriate.
+    if not is_syslog_supported():
+        return None
     # Provide defaults for omitted arguments.
     programname = programname or find_program_name()
     logger = logger or logging.getLogger()
@@ -214,6 +225,57 @@ def find_syslog_address():
         return LOG_DEVICE_UNIX
     else:
         return 'localhost', logging.handlers.SYSLOG_UDP_PORT
+
+
+def is_syslog_supported():
+    """
+    Determine whether system logging is supported.
+
+    :returns:
+
+        :data:`True` if system logging is supported and can be enabled,
+        :data:`False` if system logging is not supported or there are good
+        reasons for not enabling it.
+
+    The decision making process here is as follows:
+
+    Override
+     If the environment variable ``$COLOREDLOGS_SYSLOG`` is set it is evaluated
+     using :func:`~humanfriendly.coerce_boolean()` and the resulting value
+     overrides the platform detection discussed below, this allows users to
+     override the decision making process if they disagree / know better.
+
+    Linux / UNIX
+     On systems that are not Windows or MacOS (see below) we assume UNIX which
+     means either syslog is available or sending a bunch of UDP packets to
+     nowhere won't hurt anyone...
+
+    Microsoft Windows
+     Over the years I've had multiple reports of :pypi:`coloredlogs` spewing
+     extremely verbose errno 10057 warning messages to the console (once for
+     each log message I suppose) so I now assume it a default that
+     "syslog-style system logging" is not generally available on Windows.
+
+    Apple MacOS
+     There's cPython issue `#38780`_ which seems to result in a fatal exception
+     when the Python interpreter shuts down. This is (way) worse than not
+     having system logging enabled. The error message mentioned in `#38780`_
+     has actually been following me around for years now, see for example:
+
+     - https://github.com/xolox/python-rotate-backups/issues/9 mentions Docker
+       images implying Linux, so not strictly the same as `#38780`_.
+
+     - https://github.com/xolox/python-npm-accel/issues/4 is definitely related
+       to `#38780`_ and is what eventually prompted me to add the
+       :func:`is_syslog_supported()` logic.
+
+    .. _#38780: https://bugs.python.org/issue38780
+    """
+    override = os.environ.get("COLOREDLOGS_SYSLOG")
+    if override is not None:
+        return coerce_boolean(override)
+    else:
+        return not (on_windows() or on_macos())
 
 
 def match_syslog_handler(handler):